Re-structured code and added some comments
